sechallenge/requirements.txt
Django==1.7.1
django-uuidfield==0.5.0
